Rename files and componants
