+++ frontend/src/components/organisms/form-sections/HeadSection/CallToWriteMessageNotification.js
@@ -1,8 +1,8 @@
 import { useContext } from 'react';
 import HighlightWithButton from '../../../molecules/HighlightWithButton';
-import { OpenMessagePromptContext } from '../../../templates/Form/FormSectionsContainer';
+import { OpenMessagePromptContext } from '../../../templates/Form/OpenMessagePromptContextProvider';
 
-export const MessageFeed = ({ title }) => {
+export const CallToWriteMessageNotification = ({ title }) => {
   const { onClick } = useContext(OpenMessagePromptContext);
 
   return (
@@ -14,4 +14,4 @@ export const MessageFeed = ({ title }) => {
   );
 };
 
-export default MessageFeed;
+export default CallToWriteMessageNotification;
